+++ packages/core/kind/tests/computed-specs.js
@@ -15,7 +15,7 @@ describe('computed', () => {
 		z: 4
 	};
 
-	it('should add new props to updated object', function () {
+	test('should add new props to updated object', () => {
 		const props = {
 			value: true
 		};
@@ -28,10 +28,10 @@ describe('computed', () => {
 		const expected = 1;
 		const actual = updated.count;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should overwrite properties with computed values', function () {
+	test('should overwrite properties with computed values', () => {
 		const props = {
 			value: true,
 			count: 2
@@ -45,41 +45,44 @@ describe('computed', () => {
 		const expected = 1;
 		const actual = updated.count;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should not leak updated prop values into other computed props', function () {
-		const props = {
-			count: 1
-		};
-		const cfg = {
-			value: ({count}) => count + 5,
-			count: ({count}) => count + 1
-		};
+	test(
+		'should not leak updated prop values into other computed props',
+		() => {
+			const props = {
+				count: 1
+			};
+			const cfg = {
+				value: ({count}) => count + 5,
+				count: ({count}) => count + 1
+			};
 
-		const updated = computed(cfg, props);
+			const updated = computed(cfg, props);
 
-		const expected = 6;
-		const actual = updated.value;
-		expect(actual).to.equal(expected);
-	});
+			const expected = 6;
+			const actual = updated.value;
+			expect(actual).toBe(expected);
+		}
+	);
 
-	it('should work with its documented example - sum', function () {
+	test('should work with its documented example - sum', () => {
 		const updated = computed(exampleCfg, exampleProps);
 
 		const expected = 9;
 		const actual = updated.sum;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should work with its documented example - product', function () {
+	test('should work with its documented example - product', () => {
 		const updated = computed(exampleCfg, exampleProps);
 
 		const expected = 24;
 		const actual = updated.product;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
 });