+++ packages/moonstone/IncrementSlider/IncrementSlider.js
@@ -89,6 +89,15 @@ const IncrementSliderBase = kind({
 		'aria-valuetext': PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
 
 		/**
+		 * Disables voice control.
+		 *
+		 * @type {Boolean}
+		 * @memberof moonstone/IncrementSlider.IncrementSliderBase.prototype
+		 * @public
+		 */
+		'data-webos-voice-disabled': PropTypes.bool,
+
+		/**
 		 * The `data-webos-voice-group-label` for the IconButton of IncrementSlider.
 		 *
 		 * @type {String}
@@ -464,6 +473,7 @@ const IncrementSliderBase = kind({
 
 	render: ({active,
 		'aria-hidden': ariaHidden,
+		'data-webos-voice-disabled': voiceDisabled,
 		'data-webos-voice-group-label': voiceGroupLabel,
 		backgroundProgress,
 		css,
@@ -510,6 +520,7 @@ const IncrementSliderBase = kind({
 					aria-hidden={ariaHidden}
 					aria-label={decrementAriaLabel}
 					className={css.decrementButton}
+					data-webos-voice-disabled={voiceDisabled}
 					data-webos-voice-group-label={voiceGroupLabel}
 					disabled={decrementDisabled}
 					onTap={onDecrement}
@@ -547,6 +558,7 @@ const IncrementSliderBase = kind({
 					aria-hidden={ariaHidden}
 					aria-label={incrementAriaLabel}
 					className={css.incrementButton}
+					data-webos-voice-disabled={voiceDisabled}
 					data-webos-voice-group-label={voiceGroupLabel}
 					disabled={incrementDisabled}
 					onTap={onIncrement}