+++ packages/ui/Group/tests/Group-specs.js
@@ -1,15 +1,14 @@
 /* globals describe, it, expect */
 
 import React from 'react';
-import sinon from 'sinon';
 import {mount, shallow} from 'enzyme';
 import {GroupBase} from '../Group';
 
 describe('Group', () => {
 	const stringItems = ['One', 'Two', 'Three'];
 
-	it('should call handler with selected on select', function () {
-		const handleClick = sinon.spy();
+	test('should call handler with selected on select', () => {
+		const handleClick = jest.fn();
 		const subject = mount(
 			<GroupBase childComponent="div" onSelect={handleClick}>
 				{stringItems}
@@ -20,13 +19,13 @@ describe('Group', () => {
 		subject.find('div').at(selected).simulate('click', {});
 
 		const expected = selected;
-		const actual = handleClick.firstCall.args[0].selected;
+		const actual = handleClick.mock.calls[0][0].selected;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should call handler with data on select', function () {
-		const handleClick = sinon.spy();
+	test('should call handler with data on select', () => {
+		const handleClick = jest.fn();
 		const subject = mount(
 			<GroupBase childComponent="div" onSelect={handleClick}>
 				{stringItems}
@@ -37,30 +36,33 @@ describe('Group', () => {
 		subject.find('div').at(selected).simulate('click', {});
 
 		const expected = stringItems[selected];
-		const actual = handleClick.firstCall.args[0].data;
+		const actual = handleClick.mock.calls[0][0].data;
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should call handler on move when childSelect="onMouseMove"', function () {
-		const handleClick = sinon.spy();
-		const subject = mount(
-			<GroupBase childComponent="div" childSelect="onMouseMove" onSelect={handleClick}>
-				{stringItems}
-			</GroupBase>
-		);
+	test(
+		'should call handler on move when childSelect="onMouseMove"',
+		() => {
+			const handleClick = jest.fn();
+			const subject = mount(
+				<GroupBase childComponent="div" childSelect="onMouseMove" onSelect={handleClick}>
+					{stringItems}
+				</GroupBase>
+			);
 
-		subject.find('div').at(0).simulate('mousemove', {});
+			subject.find('div').at(0).simulate('mousemove', {});
 
-		const expected = true;
-		const actual = handleClick.called;
+			const expected = 1;
+			const actual = handleClick.mock.calls.length;
 
-		expect(actual).to.equal(expected);
-	});
+			expect(actual).toBe(expected);
+		}
+	);
 
-	it('should select the third item with selected=2', function () {
+	test('should select the third item with selected=2', () => {
 		const selected = 2;
-		const handleClick = sinon.spy();
+		const handleClick = jest.fn();
 		const subject = mount(
 			<GroupBase childComponent="div" selected={selected} onSelect={handleClick}>
 				{stringItems}
@@ -70,12 +72,12 @@ describe('Group', () => {
 		const expected = true;
 		const actual = subject.find('div').at(selected).prop('data-selected');
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should set {data-active} on the first item', function () {
+	test('should set {data-active} on the first item', () => {
 		const selected = 0;
-		const handleClick = sinon.spy();
+		const handleClick = jest.fn();
 		const subject = mount(
 			<GroupBase childComponent="div" selected={selected} selectedProp="data-active" onSelect={handleClick}>
 				{stringItems}
@@ -85,12 +87,12 @@ describe('Group', () => {
 		const expected = true;
 		const actual = subject.find('div').at(selected).prop('data-active');
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should set {children} to be the item by default', function () {
+	test('should set {children} to be the item by default', () => {
 		const selected = 0;
-		const handleClick = sinon.spy();
+		const handleClick = jest.fn();
 		const subject = mount(
 			<GroupBase childComponent="div" onSelect={handleClick}>
 				{stringItems}
@@ -100,12 +102,12 @@ describe('Group', () => {
 		const expected = stringItems[selected];
 		const actual = subject.find('div').at(selected).prop('children');
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should set {data-child} to be the item', function () {
+	test('should set {data-child} to be the item', () => {
 		const selected = 0;
-		const handleClick = sinon.spy();
+		const handleClick = jest.fn();
 		const subject = mount(
 			<GroupBase childComponent="div" childProp="data-child" onSelect={handleClick}>
 				{stringItems}
@@ -115,10 +117,10 @@ describe('Group', () => {
 		const expected = stringItems[selected];
 		const actual = subject.find('div').at(selected).prop('data-child');
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should set aria-multiselectable when select="multiple"', function () {
+	test('should set aria-multiselectable when select="multiple"', () => {
 		const subject = shallow(
 			<GroupBase childComponent="div" select="multiple">
 				{stringItems}
@@ -128,10 +130,10 @@ describe('Group', () => {
 		const expected = true;
 		const actual = subject.prop('aria-multiselectable');
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
-	it('should set role to group', function () {
+	test('should set role to group', () => {
 		const subject = shallow(
 			<GroupBase childComponent="div" select="multiple">
 				{stringItems}
@@ -141,7 +143,7 @@ describe('Group', () => {
 		const expected = 'group';
 		const actual = subject.prop('role');
 
-		expect(actual).to.equal(expected);
+		expect(actual).toBe(expected);
 	});
 
 });