+++ packages/sampler/CHANGELOG.md
@@ -2,6 +2,26 @@
 
 The following is a curated list of changes in the Enact Sampler, newest changes on the top.
 
+## [unreleased]
+
+### Changed
+
+- `moonstone/ExpandableInput` and `moonstone/Input` knob to select input type
+
+## [2.2.8] - 2018-12-06
+
+No significant changes.
+
+## [2.2.7] - 2018-11-21
+
+No significant changes.
+
+## [2.2.6] - 2018-11-15
+
+### Fixed
+
+- Sampler to allow selecting the device locale
+
 ## [2.2.5] - 2018-11-05
 
 No significant changes.